// libs/features/collection/billing/src/lib/shared/models/mof-itemized-contribution-filter.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
export class MofItemizedContributionFilter {
  minAnnuityAmount: number = undefined;
  maxAnnuityAmount: number = undefined;
  minOhAmount: number = undefined;
  maxOhAmount: number = undefined;
  minUiAmount: number = undefined;
  maxUiAmount: number = undefined;
  minTotalAmount: number = undefined;
  maxTotalAmount: number = undefined;
  adjustmentIndicator: string = undefined;
}